qcommon: fix travis build

+++ src/qcommon/q_shared.h
@@ -574,7 +574,7 @@ typedef enum
 #define Hunk_Alloc(size, preference)              Hunk_AllocDebug(size, preference, # size, __FILE__, __LINE__)
 void *Hunk_AllocDebug(unsigned int size, ha_pref preference, char *label, char *file, int line);
 #else
-void *Hunk_Alloc(size_t size, ha_pref preference);
+void *Hunk_Alloc(unsigned int size, ha_pref preference);
 #endif
 
 #define Com_Memset memset
